requirements.txt
lxml>=5.2,<6
Pillow>=10.0,<11
